requirements.txt
# Core dependencies
google-generativeai>=0.8.0
python-dotenv==1.0.0
requests>=2.31.0

# File system access
fsspec>=2024.6.0,<2025.0.0
s3fs==2024.9.0
gcsfs==2024.9.0.post1
boto3>=1.35.0,<1.36.0
botocore>=1.35.0,<1.36.0

# Type hints and utilities
typing-extensions>=4.9.0
types-requests>=2.31.0.20240311

# Optional: Only needed if we add image processing features
# Pillow>=10.0.0,<11.0.0
